Break up separator index file to reduce file size.
